adc: fix segmentation fault when exiting the application
'adc_sampling_cb_data' struct is a local variable of the 'main' function.
Sometimes, the ADC callback accesses to it once the main function has exited
(when it is out of scope), causing a segmentation fault.
To solve this issue, adc_sampling_cb_data' struct is now a global variable.

Signed-off-by: Francisco Gil <francisco.gilmartinez@digi.com>