apix-pwm-example: make a more visual PWM change on LED
When running this example on an LED-connected GPIO, the delay
of 1 second is too large, and the selected values not very good
to notice the difference in brightness.

Use a ramp array of values with a more visual difference when
running on an LED and shorten the wait between changes.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>