connectcore-demo-example: fix signal handling for DEY-4.0
With the update to DEY-4.0, the signal handling in the "demoserver.py" Python script
stopped working. The call to "event.wait()" caused the application to stop receiving
signals. This commit fixes the issue by replacing the event lock mechanism with the
"signal.sigwait()" call, which stops the thread execution until any specified signal
is received.

Signed-off-by: David Escalona <david.escalona@digi.com>