connectcore-demo-example: limit demo to use only native BT adapter
By default, when retrieving the Bluetooth interface to use in the demo, the service
first tries with the native Bluetooth adapter and, in the case it fails, it looks
for connected XBee devices with Bluetooth capabilities. This is causing the demo to
lock the XBee port if the native Bluetooth interface is not present or fails to
initialize. Avoid this behavior by limiting the Bluetooth service to use only
the native Bluetooth interface for this demo.

Signed-off-by: David Escalona <david.escalona@digi.com>