connectcore-demo-example: remove "#top" href from sidebar "collapse" button
This href causes issues when running the demo on a target with WebKit's
non-HTTP(S) process swap enabled. When clicking the "collapse" button, the
sidebar collapses, but the page is loaded again due to WebKit's process swap,
and the sidebar reverts back to its original state.

Removing the href fixes this, and the sidebar collapse functionality still
works as expected.

https://onedigi.atlassian.net/browse/DEL-9498

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>